--HG--
branch : gamification

+++ app/protected/modules/users/tests/unit/walkthrough/UsersRegularUserExportWalkthroughTest.php
@@ -27,7 +27,7 @@
     /**
      * Export module walkthrough tests.
      */
-    class UsersSuperUserExportWalkthroughTest extends ZurmoWalkthroughBaseTest
+    class UsersRegularUserExportWalkthroughTest extends ZurmoRegularUserWalkthroughBaseTest
     {
         protected static $asynchronusTreshold;
 
@@ -40,6 +40,7 @@
 
             self::$asynchronusTreshold = ExportModule::$asynchronusTreshold;
             ExportModule::$asynchronusTreshold = 5;
+            ReadPermissionsOptimizationUtil::rebuild();
         }
 
         public static function tearDownAfterClass()
@@ -54,12 +55,14 @@
         public function testDownloadDefaultControllerActions()
         {
             $super = $this->logoutCurrentUserLoginNewUserAndGetByUsername('super');
+            $nobody = User::getByUsername('nobody');
             $users = User::getAll();
+
             if (count($users))
             {
                 foreach ($users as $user)
                 {
-                    if ($user->id !== $super->id)
+                    if (($user->id !== $super->id) && ($user->id !== $nobody->id))
                     {
                         $user->delete();
                     }
@@ -70,8 +73,22 @@
             {
                 $users[] = UserTestHelper::createBasicUser('aUser' . $i);
             }
+            // Check if access is denied if user doesn't have access privileges at all to export actions
+            Yii::app()->user->userModel = User::getByUsername('nobody');
+            $nobody = $this->logoutCurrentUserLoginNewUserAndGetByUsername('nobody');
+            $this->runControllerShouldResultInAccessFailureAndGetContent('users/default/list');
 
+            // Check if user have access to module action, but not to export action
+            // Now test peon with elevated rights to accounts
+            $nobody->setRight('UsersModule', UsersModule::RIGHT_ACCESS_USERS);
+            $nobody->setRight('ExportModule', ExportModule::RIGHT_ACCESS_EXPORT);
+            $this->assertTrue($nobody->save());
+
+            // Check if access is denied if user doesn't have access privileges at all to export actions
+            $nobody = $this->logoutCurrentUserLoginNewUserAndGetByUsername('nobody');
+            Yii::app()->user->userModel = User::getByUsername('nobody');
             $this->runControllerWithNoExceptionsAndGetContent('users/default/list');
+
             $this->setGetArray(array(
                 'User_page' => '1',
                 'export' => '',
@@ -84,28 +101,28 @@
 
             $this->setGetArray(array(
                 'UsersSearchForm' => array(
-                    'anyMixedAttributesScope' => array(0 => 'All'),
+                    'anyMixedAttributesScope' => array(0 =>'All'),
                     'anyMixedAttributes'      => '',
                 ),
                 'User_page'      => '1',
                 'export'         => '',
                 'ajax'           => '',
                 'selectAll' => '1',
-                'selectedIds' => '')
+                'selectedIds' => '',)
             );
             $response = $this->runControllerWithExitExceptionAndGetContent('users/default/export');
             $this->assertEquals('Testing download.', $response);
 
             $this->setGetArray(array(
                 'UsersSearchForm' => array(
-                    'anyMixedAttributesScope' => array(0 => 'All'),
+                    'anyMixedAttributesScope' => array(0 =>'All'),
                     'anyMixedAttributes'      => '',
                 ),
                 'User_page'   => '1',
                 'export'         => '',
                 'ajax'           => '',
                 'selectAll' => '',
-                'selectedIds' => "{$users[0]->id}, {$users[1]->id}")
+                'selectedIds' => "{$users[0]->id},{$users[1]->id}")
             );
             $response = $this->runControllerWithExitExceptionAndGetContent('users/default/export');
             $this->assertEquals('Testing download.', $response);
@@ -113,114 +130,17 @@
             // No mathces
             $this->setGetArray(array(
                 'UsersSearchForm' => array(
-                    'anyMixedAttributesScope' => array(0 => 'All'),
+                    'anyMixedAttributesScope' => array(0 =>'All'),
                     'anyMixedAttributes'      => 'missingOne',
                 ),
                 'User_page' => '1',
                 'export'       => '',
                 'ajax'         => '',
                 'selectAll' => '1',
-                'selectedIds' => '')
+                'selectedIds' => '',)
             );
             $response = $this->runControllerWithRedirectExceptionAndGetUrl('users/default/export');
             $this->assertTrue(strstr($response, 'users/default/index') !== false);
-        }
-
-        /**
-        * Walkthrough test for synchronous download
-        */
-        public function testAsynchronousDownloadDefaultControllerActions()
-        {
-            $super = $this->logoutCurrentUserLoginNewUserAndGetByUsername('super');
-
-            $users = User::getAll();
-            if (count($users))
-            {
-                foreach ($users as $user)
-                {
-                    if ($user->id !== $super->id)
-                    {
-                        $user->delete();
-                    }
-                }
-            }
-            $users = array();
-            for ($i = 0; $i <= (ExportModule::$asynchronusTreshold + 1); $i++)
-            {
-                $users[] = UserTestHelper::createBasicUser('aUser' . $i);
-            }
-
-            $this->setGetArray(array(
-                'User_page' => '1',
-                'export' => '',
-                'ajax' => '',
-                'selectAll' => '1',
-                'selectedIds' => '')
-            );
-            $this->runControllerWithRedirectExceptionAndGetUrl('users/default/export');
-
-            // Start background job
-            $job = new ExportJob();
-            $this->assertTrue($job->run());
-
-            $exportItems = ExportItem::getAll();
-            $this->assertEquals(1, count($exportItems));
-            $fileModel = $exportItems[0]->exportFileModel;
-            $this->assertEquals(1, $exportItems[0]->isCompleted);
-            $this->assertEquals('csv', $exportItems[0]->exportFileType);
-            $this->assertEquals('users', $exportItems[0]->exportFileName);
-            $this->assertTrue($fileModel instanceOf ExportFileModel);
-
-            $this->assertEquals(1, count(Notification::getAll()));
-            $this->assertEquals(1, count(NotificationMessage::getAll()));
-
-            // Check export job, when many ids are selected.
-            // This will probably never happen, but we need test for this case too.
-            $notificationsBeforeCount        = count(Notification::getAll());
-            $notificationMessagesBeforeCount = count(NotificationMessage::getAll());
-
-            // Now test case when multiple ids are selected
-            $exportItems = ExportItem::getAll();
-            if (count($exportItems))
-            {
-                foreach ($exportItems as $exportItem)
-                {
-                    $exportItem->delete();
-                }
-            }
-
-            $selectedIds = "";
-            foreach ($users as $user)
-            {
-                $selectedIds .= $user->id . ","; // Not Coding Standard
-            }
-            $this->setGetArray(array(
-                'UsersSearchForm' => array(
-                    'anyMixedAttributesScope' => array(0 => 'All'),
-                    'anyMixedAttributes'      => '',
-                ),
-                'User_page'   => '1',
-                'export'         => '',
-                'ajax'           => '',
-                'selectAll' => '',
-                'selectedIds' => "$selectedIds")
-            );
-
-            $this->runControllerWithRedirectExceptionAndGetUrl('users/default/export');
-            // Start background job
-            $job = new ExportJob();
-            $this->assertTrue($job->run());
-
-            $exportItems = ExportItem::getAll();
-            $this->assertEquals(1, count($exportItems));
-            $fileModel = $exportItems[0]->exportFileModel;
-            $this->assertEquals(1, $exportItems[0]->isCompleted);
-            $this->assertEquals('csv', $exportItems[0]->exportFileType);
-            $this->assertEquals('users', $exportItems[0]->exportFileName);
-            $this->assertTrue($fileModel instanceOf ExportFileModel);
-
-            $this->assertEquals($notificationsBeforeCount + 1, count(Notification::getAll()));
-            $this->assertEquals($notificationMessagesBeforeCount + 1, count(NotificationMessage::getAll()));
         }
     }
 ?>
